reorganization of DAQ code, add sequencer tutorial
