+++ .gitignore
@@ -1,6 +1,11 @@
 ### ignore generated version file
 src/splendaq/_version.py
 
+### ignore data files
+*.li
+*.h5
+*.mat
+
 ### ignore macos DS_Store files
 *.DS_Store
