+++ src/splendaq/daq/__init__.py
@@ -1,2 +1,3 @@
-from ._continuous import *
-from ._trigger import *
+from ._log_data import *
+from ._offline_trigger import *
+from ._sequencer import *
