+++ src/splendaq/daq/_offline_trigger.py
